refactor: Header 컴포넌트 추상화
