+++ src/App.tsx
@@ -2,18 +2,22 @@ import React, { VFC } from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 import GlobalStyles from "./GlobalStyles";
-import Header from "./Components/Header";
+import { Header } from "./Components";
 import ProductList from "./Pages/ProductList";
 import Cart from "./Pages/Cart";
 import Order from "./Pages/Order";
 import { PATH } from "./constants/path";
 
+const NAVIGATION = [
+  { path: PATH.CART, name: "장바구니" },
+  { path: PATH.ORDER_LIST, name: "주문목록" },
+];
 
 const App: VFC = () => {
   return (
     <BrowserRouter>
       <GlobalStyles />
-      <Header />
+      <Header navigation={NAVIGATION} />
       <Switch>
         <Route exact path={PATH.HOME}>
           <ProductList />
