+++ src/Components/@shared/Header/style.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 
-import { COLOR } from "../../constants/theme";
+import { COLOR } from "../../../constants/theme";
 
 const Container = styled.header`
   background-color: ${COLOR.MAIN};
