+++ src/Components/@shared/index.ts
@@ -3,10 +3,11 @@ import { default as Animation } from "./Animation";
 import { default as Button } from "./Button";
 import { default as CheckBox } from "./CheckBox";
 import { default as Confirm } from "./Confirm";
+import { default as Header } from "./Header";
 import { default as Icon } from "./Icon";
 import { default as Modal } from "./Modal";
 import { default as PageTitle } from "./PageTitle";
 import { default as SubmitBox } from "./SubmitBox";
 import { default as TextWithHighlight } from "./TextWithHighlight";
 
-export { Animation, Button, CheckBox, Confirm, Icon, Modal, PageTitle, SubmitBox, TextWithHighlight };
+export { Animation, Button, CheckBox, Confirm, Header, Icon, Modal, PageTitle, SubmitBox, TextWithHighlight };
