+++ src/Components/index.ts
@@ -1,14 +1,13 @@
-import { Animation, Button, CheckBox, Confirm, Modal, PageTitle, SubmitBox } from "./@shared";
+import { Animation, Button, CheckBox, Confirm, Header, Modal, PageTitle, SubmitBox } from "./@shared";
 
 import { default as CartAnimation } from "./CartAnimation";
 import { default as CartIcon } from "./CartIcon";
 import { default as CartItem } from "./CartItem";
-import { default as Header } from "./Header";
 import { default as OrderList } from "./OrderList";
 import { default as OrderProductItem } from "./OrderProductItem";
 import { default as Portal } from "./Portal";
 import { default as Product } from "./Product";
 import { default as ProductImage } from "./ProductImage";
 
-export { Animation, Button, CheckBox, Confirm, Modal, PageTitle, SubmitBox };
-export { CartAnimation, CartItem, Header, CartIcon, OrderList, OrderProductItem, Portal, Product, ProductImage };
+export { Animation, Button, CheckBox, Confirm, Header, Modal, PageTitle, SubmitBox };
+export { CartAnimation, CartItem, CartIcon, OrderList, OrderProductItem, Portal, Product, ProductImage };
